Agregué una clase ManejadorUsuario (Ejercicio 1) y puse en el main el test unitario para probarlo.

+++ Producto.cs
@@ -19,7 +19,7 @@ namespace Proyecto_final
         public string Descripcion { get => descripcion; set => descripcion = value; }
         public decimal Costo { get => costo; set => costo = value; }
         public decimal PrecioVenta { get => precioVenta; set => precioVenta = value; }
-        public long Stock { get => stock; set => stock = value; }
+        public int Stock { get => stock; set => stock = value; }
         public long IdUsuario { get => idUsuario; set => idUsuario = value; }
     }
 }
